Print sky and sky standard deviation added.
In order to do this, I had to build a mask array
separately and using that mask array, I was able
easily find the average and standard deviation of
the sky. I also had to change the way a blank mask
is printed, compared to before.

The functions in ui.c were also re-ordered (for
easier reading). Compatible with my more recent
approach for this file in other programs.