Check default parameters in `./configure`
Until now the default values from `defaults.ac` would be directly
written into `config.h`. But now, they are checked and if there is a
problem (for example a parameter not being defined, or not being a
number or not being in the right range) `./configure` will halt and
inform the user.